fix(ci): remove unnecessary ifa_flags cast; enable ether-nd in ndn-fwd
- iface.rs: remove 'as u32' from ifa_flags — ifa_flags is already u32 in
  the current libc version (clippy::unnecessary_cast on CI's Rust 1.94)
- ndn-fwd/Cargo.toml: add ether-nd to ndn-discovery features; EtherNeighbor
  Discovery is gated behind this feature in ndn-discovery and was not being
  included in the ndn-fwd binary build, causing E0433 in the Docker build

+++ crates/faces/ndn-faces/src/iface.rs
@@ -108,7 +108,7 @@ fn list_interfaces_unix() -> Vec<InterfaceInfo> {
                 .to_string_lossy()
                 .into_owned();
 
-            let flags = (*ifa).ifa_flags as u32;
+            let flags = (*ifa).ifa_flags;
             let is_up =
                 flags & (libc::IFF_UP as u32) != 0 && flags & (libc::IFF_RUNNING as u32) != 0;
             let is_multicast = flags & (libc::IFF_MULTICAST as u32) != 0;
